[Refactoring ⚙️] Type 파일 제거
