--- src/Components/Navbar/Navbar.Types.ts
@@ -1,5 +1,0 @@
-export interface NavProfileActionsProps {
-  action: "home" | "search" | "write";
-  icon: string;
-  path: string;
-}
