--- src/Components/Navbar/components/Menubar/NavMenubar.Types.ts
@@ -1,8 +1,0 @@
-import { IconType } from "react-icons";
-
-export interface NavMenubarItemProps {
-  icon: IconType;
-  title: string;
-  category: string;
-  postCount: number;
-}
