--- src/Components/Navbar/components/NavbarFooter/components/ContactMeButton.Types.ts
@@ -1,6 +1,0 @@
-import React from "react";
-
-export interface ContactMeButtonProps {
-  url: string;
-  Icon: React.ReactNode;
-}
